Source code synced with Eugene game development kit 8.7.8

+++ source/neongdk.cpp
@@ -1829,7 +1829,7 @@ namespace NEONGDK
    return this->is_software_render();
   }
 
-  unsigned long int Screen::get_color() const
+  unsigned int Screen::get_color() const
   {
    return this->get_depth();
   }
